Move Material Editor Basic test from sandbox to main suite (#6794)
* moving material editor test from sandbox to main suite


* define TEST_DIRECTORY and add logging
